Automating Physiological Data Collection from Human Reactions to Environmental Stimuli/Code/frontend/requirements.txt
tensorflow==2.8.0
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pandas==1.2.3
python-dateutil==2.8.1
pytz==2021.1
retrying==1.3.3
six==1.15.0
threadpoolctl==2.1.0
Werkzeug==1.0.1
Pillow==8.2.0
opencv-python==4.5.5.64